+++ packages/rexical/src/plugins/CodeActionMenuPlugin/index.tsx
@@ -167,7 +167,7 @@ function getMouseInfo(event: MouseEvent): {
 
   if (isHTMLElement(target)) {
     const codeDOMNode = target.closest<HTMLElement>(
-      'code.PlaygroundEditorTheme__code',
+      'code.nim-code',
     );
     const isOutside = !(
       codeDOMNode ||